requirements.txt
fastapi
supabase
pydantic
python-dotenv
httpx
uvicorn
pydantic-ai
pydantic-ai-slim
pydantic-graph
langfuse
logfire-api
devtools
